adding new page for pdf feature
